+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Release 1.0.3
+
+**Date**: `2024-01-26`
+
+## Changes
+
+- Added Absolute Humidity calculated sensor
+- Added Freezing Altitude calculated sensor
+- Added Visibility calculated sensor
+
 # Release 1.0.2
 
 **Date**: `2024-01-26`